Added Features - Added feature dependencies - Noisy CANBus - Feature UDS bringup on BMSW and BMSB. Added compatibility to stw and bl. - Added BMSW and BMSB faults. - Added balancing feature. - 10kHz Task - BMSW high frequency task feature. - BMSW CANRX and CANTX SWI feature. CANRX SWI does not work with UDS.
YAMCAN Improvement
- Unrecorderded and unscheduled CAN message.
- Network definition file generation. UDS ID's determined at NetworkGen. Rx messages for specific duplicate nodes in NetowrkGen.
- CAN ID clash between BMS Boss UDS and Charger
- STW UDS messages.
- Add Brusa message id's
        - Disable BMSB UDS because of ID clash with charger

Known Issues
- CANRX SWI on BMSW doesnt function with UDS.
- BMSB current sense in savvycan issue.

+++ components/bms_worker/include/Utility.h
@@ -88,8 +88,8 @@ static inline uint16_t u32CountLeadingZeroes(uint32_t x)
 #define FLAG_GET_MASK(flag)       (1U << ((uint16_t)flag % FLAG_bits_each))
 
 #define FLAG_create(name, size) uint16_t(name)[WORDS_FROM_COUNT(size)]
-#define FLAG_set(name, pos)     FLAG_GET_WORD(name, pos) |= FLAG_GET_MASK(pos)
-#define FLAG_clear(name, pos)   FLAG_GET_WORD(name, pos) &= ~FLAG_GET_MASK(pos)
+#define FLAG_set(name, pos)     FLAG_GET_WORD(name, pos) |= (uint16_t)FLAG_GET_MASK(pos)
+#define FLAG_clear(name, pos)   FLAG_GET_WORD(name, pos) &= (uint16_t)~FLAG_GET_MASK(pos)
 #define FLAG_get(name, pos)     ((bool)((FLAG_GET_WORD(name, pos) & FLAG_GET_MASK(pos)) == FLAG_GET_MASK(pos)))
 #define FLAG_assign(name, pos, value) \
  do {                                 \
